+++ src/lowntfs-3g.c
@@ -3036,7 +3036,14 @@ static void ntfs_fuse_ioctl(fuse_req_t req __attribute__((unused)),
 			}
 			memcpy(buf, data, in_bufsz);
 		}
-		ret = ntfs_ioctl(ni, cmd, arg, flags, buf);
+		/*
+		 * Linux defines the request argument of ioctl() to be an
+		 * unsigned long, which fuse 2.x forwards as a signed int
+		 * into which the request sometimes does not fit.
+		 * So we must expand the value and make sure it is not
+		 * sign-extended.
+		 */
+		ret = ntfs_ioctl(ni, (unsigned int)cmd, arg, flags, buf);
 		if (ntfs_inode_close (ni))
 			set_fuse_error(&ret);
 	}
